+++ e2e/fixtures/caseData/gatekeepingNoHearingDetails.json
