+++ e2e/fixtures/caseData/issueResolution.json
