+++ e2e/fixtures/caseData/mandatorySubmissionFields.json
