+++ e2e/fixtures/caseData/mandatoryWithMultipleChildren.json
