+++ e2e/tests/gatekeeperMakesAllocationDecision_test.js
@@ -1,5 +1,5 @@
 const config = require('../config.js');
-const gatekeeping = require('../fixtures/gatekeeping.json');
+const gatekeeping = require('../fixtures/caseData/gatekeeping.json');
 
 let caseId;
