+++ e2e/tests/localAuthorityMaintainsCaseAfterSubmission_test.js
@@ -2,7 +2,7 @@ const config = require('../config.js');
 const recipients = require('../fixtures/recipients.js');
 const placementHelper = require('../helpers/placement_helper.js');
 const uploadDocumentsHelper = require('../helpers/upload_case_documents_helper.js');
-const mandatoryWithMultipleChildren = require('../fixtures/mandatoryWithMultipleChildren.json');
+const mandatoryWithMultipleChildren = require('../fixtures/caseData/mandatoryWithMultipleChildren.json');
 
 let caseId;
